neo-desktop-agent/requirements.txt
# NEO Desktop Agent - todo lo necesario para cualquier tipo de petición
# Python 3.10+
# Tras pip install -r requirements.txt ejecuta: playwright install chromium

# --- Telegram y control del escritorio ---
python-telegram-bot>=20.0
pyautogui>=0.9.50
playwright>=1.40.0
Pillow>=10.0.0

# --- Documentos (PDF, Word, Excel, presentaciones) ---
pypdf>=3.0.0
reportlab>=4.0.0
python-docx>=1.0.0
python-pptx>=0.6.0
openpyxl>=3.1.0

# --- Web: HTTP, scraping, parsing HTML ---
requests>=2.28.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
duckduckgo-search>=6.0.0

# --- Imágenes: Gemini (PDF con imagen, skill generar_imagen, libro magia) ---
google-genai>=1.0.0

# --- Imágenes y OCR ---
# pytesseract (necesita Tesseract instalado en el sistema para OCR)
pytesseract>=0.3.10

# --- Audio: transcripción de voz ---
openai-whisper>=20231117

# --- Datos y análisis (tablas, números, gráficos) ---
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0

# --- Skills: QR, estado del sistema ---
qrcode>=7.0.0
psutil>=5.9.0

# --- Utilidades: clipboard, markdown, YAML, env ---
pyperclip>=1.8.0
markdown>=3.5.0
PyYAML>=6.0
python-dotenv>=1.0.0
schedule>=1.2.0
fastapi>=0.100.0
uvicorn>=0.22.0
